Support live stream of data
This could definitely be improved in the future. At first I tried using an SSE data stream so that just one request is made from the client, and then the server continuously sends a single stream of data. This would be the optimal solution. However I was having lots of trouble getting this working, so for now I'm using a setInterval to send multiple requests on an interval and receive multiple responses. This will have a lot of limitations (namely speed) and I would like to revisit using SSE in the future. However, it'll work for now.